requirements.txt
streamlit>=1.29.0
pandas>=2.1.3
openpyxl>=3.1.2
pymupdf>=1.26.0
pytesseract>=0.3.13
pdf2image>=1.17.0
pillow>=11.0.0
